Added Clement 2018(19) initial conditions to example folder
This example has an extra ice giant in it (in keeping with the Clement 2018 model) which I haven't previously included. Its orbital elements are a random mix of the orbital elements of Uranus and Neptune except that it has a semimajor axis of 35AU (again, in keeping with the Clement 2018 model). I am not sure how stable this system is, so I may test it and remove it if it doesn't seem like a good fit.